✨ refactor: update systems and core logic
