+++ packages/core-logic/src/index.ts
@@ -1,1 +1,3 @@
 export * from './agent/Orchestrator';
+---                   typescript
+                    export * from './agent/Orchestrator';
