+++ shared/index.ts
@@ -1,2 +1,2 @@
-export * from "./types/weather";
-export * from "./types/are-engine-box";
+export * from "./types/Weather";
+export * from "./types/AreEngineBox";
